backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Pillow==10.4.0
pytest==7.4.4
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.4.4
flake8==7.0.0
reportlab==4.2.2
django-filter==24.3